+++ rover-server/internal/controller/apispecification_test.go
@@ -6,20 +6,44 @@ package controller
 
 import (
 	"bytes"
+	"context"
 	"encoding/json"
+	"io"
 	"net/http"
 	"net/http/httptest"
 
 	"github.com/gkampitakis/go-snaps/match"
 	. "github.com/onsi/ginkgo/v2"
+	"github.com/stretchr/testify/mock"
 
+	fileApi "github.com/telekom/controlplane/file-manager/api"
 	"github.com/telekom/controlplane/rover-server/internal/api"
 )
 
 var _ = Describe("ApiSpecification Controller", func() {
+
+	specV3 := `
+openapi: "3.0.0"
+info:
+  version: "1.0.0"
+  title: "Rover API"
+servers: 
+  - url: "http://rover-api.com/eni/distr/v1"
+`
+
 	Context("Get ApiSpecification resource", func() {
 		It("should return the ApiSpecification successfully", func() {
-			req := httptest.NewRequest(http.MethodGet, "/apispecifications/eni--hyperion--apispec-sample", nil)
+			mockFileManager.EXPECT().DownloadFile(mock.Anything, mock.Anything, mock.Anything).
+				RunAndReturn(func(_ context.Context, _ string, w io.Writer) (*fileApi.FileDownloadResponse, error) {
+
+					w.Write([]byte(specV3))
+
+					return &fileApi.FileDownloadResponse{
+						FileHash:    "randomHash",
+						ContentType: "application/yaml",
+					}, nil
+				})
+			req := httptest.NewRequest(http.MethodGet, "/apispecifications/eni--hyperion--eni-distr-v1", nil)
 			responseGroup, err := ExecuteRequest(req, groupToken)
 			ExpectStatusWithBody(responseGroup, err, http.StatusOK, "application/json", match.Any("status.time"))
 		})
@@ -31,7 +55,7 @@ var _ = Describe("ApiSpecification Controller", func() {
 		})
 
 		It("should fail to get an ApiSpecification from a different team", func() {
-			req := httptest.NewRequest(http.MethodGet, "/apispecifications/other--team--apispec-sample", nil)
+			req := httptest.NewRequest(http.MethodGet, "/apispecifications/other--team--eni-distr-v1", nil)
 			responseGroup, err := ExecuteRequest(req, groupToken)
 			ExpectStatusWithBody(responseGroup, err, http.StatusForbidden, "application/problem+json")
 		})
@@ -39,6 +63,17 @@ var _ = Describe("ApiSpecification Controller", func() {
 
 	Context("GetAll ApiSpecifications resource", func() {
 		It("should return all ApiSpecifications successfully", func() {
+			mockFileManager.EXPECT().DownloadFile(mock.Anything, mock.Anything, mock.Anything).
+				RunAndReturn(func(_ context.Context, _ string, w io.Writer) (*fileApi.FileDownloadResponse, error) {
+
+					w.Write([]byte(specV3))
+
+					return &fileApi.FileDownloadResponse{
+						FileHash:    "randomHash",
+						ContentType: "application/yaml",
+					}, nil
+				})
+
 			req := httptest.NewRequest(http.MethodGet, "/apispecifications", nil)
 			responseGroup, err := ExecuteRequest(req, groupToken)
 			ExpectStatusWithBody(responseGroup, err, http.StatusOK, "application/json", match.Any("items.0.status.time"))
@@ -59,7 +94,7 @@ var _ = Describe("ApiSpecification Controller", func() {
 
 	Context("Delete ApiSpecification resource", func() {
 		It("should delete the ApiSpecification successfully", func() {
-			req := httptest.NewRequest(http.MethodDelete, "/apispecifications/eni--hyperion--apispec-sample", nil)
+			req := httptest.NewRequest(http.MethodDelete, "/apispecifications/eni--hyperion--eni-distr-v1", nil)
 			responseGroup, err := ExecuteRequest(req, groupToken)
 			ExpectStatus(responseGroup, err, http.StatusNoContent, "")
 		})
@@ -71,7 +106,7 @@ var _ = Describe("ApiSpecification Controller", func() {
 		})
 
 		It("should fail to delete an ApiSpecification from a different team", func() {
-			req := httptest.NewRequest(http.MethodDelete, "/apispecifications/other--team--apispec-sample", nil)
+			req := httptest.NewRequest(http.MethodDelete, "/apispecifications/other--team--eni-distr-v1", nil)
 			responseGroup, err := ExecuteRequest(req, groupToken)
 			ExpectStatusWithBody(responseGroup, err, http.StatusForbidden, "application/problem+json")
 		})
@@ -79,7 +114,7 @@ var _ = Describe("ApiSpecification Controller", func() {
 
 	Context("GetStatus ApiSpecification resource", func() {
 		It("should return the status of the ApiSpecification successfully", func() {
-			req := httptest.NewRequest(http.MethodGet, "/apispecifications/eni--hyperion--apispec-sample/status", nil)
+			req := httptest.NewRequest(http.MethodGet, "/apispecifications/eni--hyperion--eni-distr-v1/status", nil)
 			responseGroup, err := ExecuteRequest(req, groupToken)
 			ExpectStatusWithBody(responseGroup, err, http.StatusOK, "application/json")
 		})
@@ -91,7 +126,7 @@ var _ = Describe("ApiSpecification Controller", func() {
 		})
 
 		It("should fail to get the status of an ApiSpecification from a different team", func() {
-			req := httptest.NewRequest(http.MethodGet, "/apispecifications/other--team--apispec-sample/status", nil)
+			req := httptest.NewRequest(http.MethodGet, "/apispecifications/other--team--eni-distr-v1/status", nil)
 			responseGroup, err := ExecuteRequest(req, groupToken)
 			ExpectStatusWithBody(responseGroup, err, http.StatusForbidden, "application/problem+json")
 		})
@@ -113,16 +148,29 @@ var _ = Describe("ApiSpecification Controller", func() {
 		It("should update the ApiSpecification successfully", func() {
 			var apiSpecification, _ = json.Marshal(api.ApiSpecificationCreateRequest{
 				Specification: map[string]interface{}{
+					"openapi": "3.0.0",
 					"info": map[string]interface{}{
-						"title":   "Rover API",
-						"version": "1.0.0",
+						"title":      "Rover API",
+						"version":    "1.0.0",
+						"x-category": "other",
+						"x-vendor":   "true",
 					},
-					"servers": map[string]interface{}{
-						"url": "http://rover-api/eni/distr/v1",
+					"servers": []map[string]interface{}{
+						{
+							"url": "http://rover-api.com/eni/distr/v1",
+						},
 					},
 				},
 			})
-			req := httptest.NewRequest(http.MethodPut, "/apispecifications/eni--hyperion--apispec-sample",
+
+			mockFileManager.EXPECT().UploadFile(mock.Anything, mock.Anything, mock.Anything, mock.Anything).Return(
+				&fileApi.FileUploadResponse{
+					FileHash:    "randomHash",
+					FileId:      "randomId",
+					ContentType: "application/yaml",
+				}, nil)
+
+			req := httptest.NewRequest(http.MethodPut, "/apispecifications/eni--hyperion--eni-distr-v1",
 				bytes.NewReader(apiSpecification))
 			responseGroup, err := ExecuteRequest(req, groupToken)
 			ExpectStatusWithBody(responseGroup, err, http.StatusAccepted, "application/json", match.Any("status.time"))
@@ -131,16 +179,21 @@ var _ = Describe("ApiSpecification Controller", func() {
 		It("should fail to update a non-existent ApiSpecification", func() {
 			var apiSpecification, _ = json.Marshal(api.ApiSpecificationCreateRequest{
 				Specification: map[string]interface{}{
+					"openapi": "3.0.0",
 					"info": map[string]interface{}{
-						"title":   "Rover API",
-						"version": "1.0.0",
+						"title":      "Rover API",
+						"version":    "1.0.0",
+						"x-category": "test",
+						"x-vendor":   "true",
 					},
-					"servers": map[string]interface{}{
-						"url": "http://rover-api/eni/distr/v1",
+					"servers": []map[string]interface{}{
+						{
+							"url": "http://rover-api.com/not/there/v1",
+						},
 					},
 				},
 			})
-			req := httptest.NewRequest(http.MethodPut, "/apispecifications/eni--hyperion--blabla",
+			req := httptest.NewRequest(http.MethodPut, "/apispecifications/eni--hyperion--not-there-v1",
 				bytes.NewReader(apiSpecification))
 			responseGroup, err := ExecuteRequest(req, groupToken)
 			ExpectStatusWithBody(responseGroup, err, http.StatusNotFound, "application/problem+json")
@@ -149,16 +202,21 @@ var _ = Describe("ApiSpecification Controller", func() {
 		It("should fail to update an ApiSpecification from a different team", func() {
 			var apiSpecification, _ = json.Marshal(api.ApiSpecificationCreateRequest{
 				Specification: map[string]interface{}{
+					"openapi": "3.0.0",
 					"info": map[string]interface{}{
-						"title":   "Rover API",
-						"version": "1.0.0",
+						"title":      "Rover API",
+						"version":    "1.0.0",
+						"x-category": "test",
+						"x-vendor":   "true",
 					},
-					"servers": map[string]interface{}{
-						"url": "http://rover-api/eni/distr/v1",
+					"servers": []map[string]interface{}{
+						{
+							"url": "http://rover-api/eni/distr-other/v1",
+						},
 					},
 				},
 			})
-			req := httptest.NewRequest(http.MethodPut, "/apispecifications/other--team--apispec-sample",
+			req := httptest.NewRequest(http.MethodPut, "/apispecifications/other--team--eni-distr-other-v1",
 				bytes.NewReader(apiSpecification))
 			responseGroup, err := ExecuteRequest(req, groupToken)
 			ExpectStatusWithBody(responseGroup, err, http.StatusForbidden, "application/problem+json")